fix bug in invoice from clipboard.
The problematic case was when the clipboard contained a string starting
with "ln", the code interpret this as a lightning invoice candidate and
prompted error when decoding failed.
The other problematic case was that invoice could be sent to breez
via multiple sources simultaniousely, e.g clipboard and link.
The flow now is more robust by giving a chance for both flows and the
first flow that succeeds wins the prompt.